Allow forum objects to both lock and unlock.
The primary purpose of this commit is to make the forum class slightly more
flexible.  Alex and I found that this flexibility is needed in our games, and
without this change we would have had to use something of a hack to work around
it.

This commit has not been tested, and so it probably does not work.