Create the conversation class.
The conversation class is part of the messaging module.  Unlike the forum,
which does not distinguish between any of its members, the conversation allows
for direct communication over a pipe.  I think this provides some nice balance.

Along with the additions to the messaging module, there were some small changes
to the networking module.  Pipes are still capable of assigning themselves ID
numbers, but I will remove that feature shortly.