Redesign the game engine to be more concrete.
This core game engine redesign was inspired by the desire to make it
more difficult to misuse and the need to make it more robust against cheaters.
Both of these goals were accomplished by allowing the game engine to implement
a more sophisticated messaging system than that supported by the forum.

This commit also includes a rearrangement and consolidation of some of the
other game engine modules.