Begin adding the concept of actor proxies.
Actor proxies provide a way for the game engine to prevent message objects
from passing raw game tokens to the actors.  In many ways, these proxies are
the opposite of token proxies.

In fact, I found the need for actor proxies while trying to integrate token
proxies into the game.  I wanted messages to be able to directly call
methods in the actors, but this made it very easy for the message object to
provide raw tokens to the actors.