Check that messages are sent by the right players.
In the previous iterations of the game engine, it would have been possible for
malicious players to send messages as the wrong player.  To address, this, the
game engine now provides a player argument to the message.check() function.
The player is associated with each pipe as the game is being set up, so this
method cannot be hacked.