Make the Token class pickle-able.
This required getting rid of the TokenProxy architecture and as a result, the
permission protections are now weaker.  Access to all methods is allowed by
default, and access is revoked on a case by case basis.  However, the system no
longer requires any metaclass programming and can be translated to C+++ much
more directly.