Add tougher restrictions to token access.
This commit implements a system for controlling token access based on method
decorators and metaclasses.  This system is stronger because it requires that
read-only methods be explicitly marked, and it also does not have any effect on
how tokens are pickled.

The unit tests supporting this new feature are present, but relatively weak.  I
need to develop a more comprehensive test set, perhaps composed of more
integration-type tests.