Remove greetings and guard the world better.
The greeting mechanism provided a way to associate a game object with a
particular actor, for the purpose of validating messages.  However, it
was a opaque and awkward, partially because it referenced game-specific
tokens from within the game engine.  This commit replaces the greeting
system with a simpler one based on ID numbers.  Each actor is assigned
an unique ID when the game starts up, and these IDs can be used to
validate messages.

This commit also adds more sophisticated safety checks to the game
world.  Token methods that aren't read-only should only be called from
messages.  Previously, the game engine could only enforce this if the
non-read-only methods were labeled with a decorator.  Any undecorated
methods were assumed to be read-only.  This was not a good approach,
because it was too easy to simply label nothing.

This commit forces the read-only methods to be labeled, and assumes all
token methods are dangerous by default.  This works by having a
metaclass effectively decorate all token methods behind the scenes.
Fortunately, metaclasses do not seem to have any effect on pickling.

This commit also creates a publisher/subscriber system for use with
messages.  In other words, Message.execute() and Message.notify() have
been removed.  The game engine will now search for methods with the
proper name (e.g. handle_create_soldier) in the world and actor objects,
and automatically invoke those.  This approach removes a lot of
boilerplate, but it also adds the potential for some more subtle bugs.
For example, if the name of a handler is misspelled, the handler will be
silently ignored.  This may be addressed in future commits.

Finally, this commit adds single player and multiplayer integration
tests.  Hopefully these tests will make future game engine development
easier and more robust.