Add helpful output to MultiplayerDebugger.
I added a logger object that intercepts all the output generated in each game
process, and displays it prefixed with the name of the process.  This makes it
much easier to see which game is outputting which message, especially when the
output is badly interleaved.  The logger can optionally copy the output stream
from each game into a different file.