Convert Pipe.receive() into a generator.
This fixes a very subtle bugs that occurs when two messages are received in the
same frame, and the second message contains a reference to a token that will be
created by the first.  The bug was caused because both messages were being
unpickled before being processed.  The second message then failed to be
unpickled because the first message had not yet been processed.  Returning an
iterator instead of a list solved the problem nicely.

This commit also adds some more documentation and a unit test specifically for
this bug.