Implement the sent message cache.
Prior to this commit, messages received by the server would be broadcast
to all clients, including the client that sent the message in the first
place.  This turned out to be problematic (in addition to being slightly
wasteful) because that client would have already executed the message
and may not be able to unpickle it anymore.  For example, if the message
removed a token, that token would no longer no longer exist on the
client and an exception would be thrown when the client tried to
unpickle the message.

This commit adds a "sent message cache" to clients that keeps track of
the most recent messages sent by the client.  Instead of attaching its
response to the message itself, the server now responds with a separate
packet.  When the client receives such a response from the server, it
pops the relevant message off its sent message cache and handles the
response as appropriate.

All tests passed, but some naming conventions still have to be updated.