Overload the rshift operator to send messages.
I'm not really sure if this is a good idea.  The argument in favor is
that sending messages is a really common thing for actors to do, and
this syntax is less cluttered.  The argument in favor is that this is an
abuse of operator overloading, and that the it makes the API harder to
use for unfamiliar users.  For comparison:

actor.send_message(CreateSoldier())
actor >> CreateSoldier()

I'm not sure how I feel, so I'm going to test drive it for a while.