Store information about the GUI in Theater.
A few key GUI objects (e.g. the window) need to be shared between the
stages that make up the theater.  Previously, this information would be
manually passed between each stage.  This approach was always kinda
clunky, but it really became a problem for the quickstart code, because
it doesn't expose the stages to the user.

An alternative was to store these key GUI objects in the theater itself.
This makes sense because all the stages are using them anyways.  It also
works well with the quickstart code, because the stages no longer need
to explicitly configure each other.  The biggest drawback is that adding
implicit references to the GUI in the theater makes that class less
conceptually pure.  However, in this case I think pragmatism wins and
the additional responsibility of the Theater class is worth the decrease
in acrobatics that users have to go through to pass GUI objects around.

Theater can now be given an arbitrary "gui" object.  This object can be
accessed by any stage, and GameStage additionally makes it available to
each of its actors.  Quickstart.main() now requires a "gui" class.  If
the class provides an on_refresh_gui() method, it will be called before
and after the game as appropriate to clear the screen and to display any
"game over" messages set at the end of the game.

I thought a lot about providing a default "gui" object in quickstart
games, but ultimately I decided that this would it would just be easier
for players to write their own.  In other words, I didn't think a
default would provide enough features to justify the learning curve.