Require actors to explicitly store the GUI.
When I added the Gui object to the engine a few commits ago, I had the
game engine automatically store a reference to it in each actor.  But as
I tried to explain how the GuiActor ends up with this reference in the
documentation for the Guess My Number demo, I came to believe that this
process was too magical.  Now the Gui object is passed as an argument to
on_setup_gui(), and each actor can do with that as it pleases.

This commit also completes the documentation for the GuiActor section of
the Guess My Number demo.