Tell the actors how many players there are.
The referee needs this information so it can start the game once it has
received all of the "CreatePlayer" messages it's expecting, but this
implementation is kinda hacky.  It will be worth thinking about this
more deeply.  In particular, how can the game indicate that there
shouldn't be more than 4 players, for example?