Make ApiUsageError easier to use directly.
Before ApiUsageError was really just a backend for ApiUsageErrorFactory.
To report an error, you would subclass ApiUsageErrorFactory in errors.py
and raise your subclass wherever you needed to.

However, I now think it would be better to directly raise ApiUsageError
in most cases.  This puts the text of the error message at the same
place the error was detected, which serves as built-in documentation.
The ApiUsageErrorFactory is still useful for errors that can be
triggered in several different ways, but ApiUsageError is better for
one-off errors.