Move pre-world safety checks from tokens to messages.
Previously, tokens could be in three states with respect to the world:
before being added, currently participating, and after being removed.
The first and last states were distinguished only for the purpose of
error checking.

However, I decided that the "before being added" checks were too onerous
because they were getting in the way of unit tests for other games.  So
I combined the first and last states into one "not participating" state.
I removed any error checking the token does on itself when it's not
participating in the world, and instead added more error checking at the
point where tokens are being sent in messages.

The new participating and not participating states are not explicitly
represented.  Instead, they are implied by the value of Token.has_world.
Token.has_id is generally not checked and is assumed to be consistent
with Token.has_world.  I'm not sure this lack of explicit representation
is a good idea.  For example, if I want to write a unit test where a
token needs to use its world attribute, do I really want error checks?
That said, there's no proper way to even give a token a world attribute
without adding it to the world, and if you have a world object anyway,
it's not so hard to lock it temporarily.

This commit also streamlines how errors are reported to the user.
Before, it wasn't uncommon for the same error to be checked for in
several different places, all with several different error messages.
This was clearly a waste of effort.  It also made testing difficult
because it was sometimes difficult to trigger subtly different error
conditions, and often these conditions could only be triggered by
improperly calling private functions.

To improve this situation, I adopted the convention of using raising
ApiUsageError when an error is noticed for the first time. This is
usually in a public API function, and Actor.send_message() is especially
a hub for these checks.  In all other cases, I simply assert that the
expected condition is still satisfied, and provide a terse explanation
saying where the error should've been caught (with an ApiUsageError) if
the condition is not satisfied.

This way, users of the library should only ever see ApiUsageErrors, but
developers of the library may also see assertion errors.