Move the Theater and Stage classes to quickstart.
The Theater and Stage classes never really belonged in the core game
engine.  They manage the main loop rather than helping with networked
communication, and the core game engine should be agnostic to the main
loop.

When I added these classes, it was for reasons of convenience vs purity.
However, now the quickstart module provides convenience via a main loop
that is loop that is separate from the core game engine and that is not
required to be all that flexible.

The quickstart module is a more natural home for the Theater and Stage
classes, but until not I had not gotten around to moving them.  In fact,
I suspect that the Theater and Stage classes are more flexible than the
quickstart main loop requires, and that this flexibility might even be
to the detriment of features like pre- and post-game GUIs.